Rename hierarchy -> hierarchical_views
